Consolidate duplicate play-ws tests
