+++ instrumentation/play-ws/play-ws-testing/src/main/groovy/PlayWsClientTestBaseBase.groovy
@@ -1,0 +1,66 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import akka.actor.ActorSystem
+import akka.stream.ActorMaterializer
+import akka.stream.ActorMaterializerSettings
+import io.opentelemetry.instrumentation.test.AgentTestTrait
+import io.opentelemetry.instrumentation.test.base.HttpClientTest
+import play.shaded.ahc.org.asynchttpclient.AsyncHttpClient
+import play.shaded.ahc.org.asynchttpclient.AsyncHttpClientConfig
+import play.shaded.ahc.org.asynchttpclient.DefaultAsyncHttpClient
+import play.shaded.ahc.org.asynchttpclient.DefaultAsyncHttpClientConfig
+import spock.lang.Shared
+
+abstract class PlayWsClientTestBaseBase extends HttpClientTest implements AgentTestTrait {
+  @Shared
+  ActorSystem system
+
+  @Shared
+  AsyncHttpClient asyncHttpClient
+
+  @Shared
+  ActorMaterializer materializer
+
+  def setupSpec() {
+    String name = "play-ws"
+    system = ActorSystem.create(name)
+    ActorMaterializerSettings settings = ActorMaterializerSettings.create(system)
+    materializer = ActorMaterializer.create(settings, system, name)
+
+    AsyncHttpClientConfig asyncHttpClientConfig =
+      new DefaultAsyncHttpClientConfig.Builder()
+        .setMaxRequestRetry(0)
+        .setShutdownQuietPeriod(0)
+        .setShutdownTimeout(0)
+        .build()
+
+    asyncHttpClient = new DefaultAsyncHttpClient(asyncHttpClientConfig)
+  }
+
+  def cleanupSpec() {
+    system?.terminate()
+  }
+
+  String expectedOperationName() {
+    return "play-ws.request"
+  }
+
+  @Override
+  boolean testCircularRedirects() {
+    return false
+  }
+
+  @Override
+  boolean testCallbackWithParent() {
+    return false
+  }
+
+  @Override
+  boolean testRemoteConnection() {
+    // TODO(anuraaga): Timeouts contain ConnectException in client span instead of TimeoutException
+    return false
+  }
+}
